- fixed debug rendering - moved prototypes to old folder - added util function for setting empty img
